requirements.txt
altair==5.3.0
numpy==1.24.4
pandas==1.4.4
Pillow==9.2.0
Pillow==10.3.0
streamlit==1.35.0
